Merge final  7.5 release  into master

See merge request QEF/q-e!2656

+++ EPW/src/wannier.f90
@@ -103,7 +103,7 @@
     !! Counter on coarse q-point grid
     INTEGER :: lrepmatw
     !! record length while reading file
-    INTEGER :: valueRSS(2)
+    INTEGER(KIND = 8) :: valueRSS(2)
     !! Return virtual and resisdent memory from system
     INTEGER :: ierr
     !! Error status
